Implemented /track & /trackoff
Two new commands
/track <username> & /trackoff

This will /whois username every 4 seconds, and once they are in-game, it will automatically search for that users game via /game gamename.

This will also play track_done.wav once the player is in-game. This is to notify the user is in-game, and GProxy is searching for their game name.

Code was provided by a previous suggestion made in 2011.
(https://github.com/w3gh/gproxyplusplus/issues/5)
• Sound notification works/tested on Windows/Mac.
• track_done.wav provided with compiled GProxy on ENT Gaming Wiki.